fixed typos
